Refactor server creation
