+++ actions/serverManagement/createServer.ts
@@ -5,60 +5,63 @@ import {
   createServerConfig,
   createServerEntry,
   createServerMember,
+  deleteServerData,
 } from "../../prisma/services/serverService";
 import * as bcrypt from "bcryptjs";
 
 type newServerData = {
-  serverName: string;
-  description?: string;
-  image?: string;
-  isProtected?: boolean;
-  password?: string;
-  explorePermission?: boolean;
-  searchPermission?: boolean;
-  joinPermission?: string;
-  settingsRightsHolders: string;
+  serverData: {
+    server_name: string;
+    description?: string;
+    image?: string;
+  };
+  configData: {
+    protected?: boolean | null;
+    password?: string | null;
+    explorable?: boolean | null;
+    searchable?: boolean | null;
+    join_permission?: string | null;
+    config_permission: string;
+  };
 };
 
 export default async function createServer(
   user_id: string,
   data: newServerData,
 ) {
-  if (data.isProtected && data.password === "") {
+  if (data.configData.protected && data.configData.password === "") {
     throw new Error("Server is set to protected but no password was given.");
   }
-  const server = await createServerEntry({
-    server_name: data.serverName,
-    description: data.description,
-    image: data.image || undefined,
-  });
+  const server = await createServerEntry(data.serverData);
 
-  const password_hash = await bcrypt.hash(String(data.password), 10);
+  const password_hash = await bcrypt.hash(String(data.configData.password), 10);
 
-  const serverConfig = await createServerConfig({
-    server_id: server.id,
-    config_permission: data.settingsRightsHolders,
-    protected: data.isProtected,
-    password_hash: data.isProtected ? password_hash : undefined,
-    explorable: data.explorePermission,
-    searchable: data.searchPermission,
-    join_permission: data.joinPermission,
-  });
+  try {
+    const serverConfig = await createServerConfig({
+      ...data.configData,
+      server_id: server.id,
+      password_hash: data.configData.protected ? password_hash : undefined,
+    });
 
-  const serverOwner = await createServerMember({
-    server_id: server.id,
-    member_id: user_id,
-    role: "admin",
-  });
+    const serverOwner = await createServerMember({
+      server_id: server.id,
+      member_id: user_id,
+      role: "admin",
+    });
 
-  const serverCharacterConfig = await createServerCharacterConfig(server.id, {
-    vitals_count: 1,
-    vitals_names: ["HP"],
-    attributes_count: 2,
-    attributes_names: ["Strength", "Dexterity"],
-    statics_count: 2,
-    statics_names: ["Stealth", "Persuasion"],
-  });
+    const serverCharacterConfig = await createServerCharacterConfig(server.id, {
+      vitals_count: 1,
+      vitals_names: ["HP"],
+      attributes_count: 2,
+      attributes_names: ["Strength", "Dexterity"],
+      statics_count: 2,
+      statics_names: ["Stealth", "Persuasion"],
+    });
+  } catch (e) {
+    await deleteServerData(server.id);
+
+    throw new Error("Server creation failed.");
+  }
 
   return server;
 }
