+++ app/(user)/server/create/page.tsx
@@ -1,7 +1,7 @@
 import Main from "@/app/_components/wrappers/PageMain";
 import { auth } from "@/auth";
 import { redirect } from "next/navigation";
-import CreateServer from "./_components/CreateServer";
+import CreateServerForm from "./_components/CreateServerForm";
 
 export default async function CreateServerPage() {
   const session = await auth();
@@ -10,7 +10,7 @@ export default async function CreateServerPage() {
 
   return (
     <Main className="ml-8 ps-12 pt-2">
-      <CreateServer userId={(session as ExtendedSession).userId} />
+      <CreateServerForm userId={(session as ExtendedSession).userId} />
     </Main>
   );
 }
